fixed: split header and body components
