+++ screens/Home/components/body/BodyComponent.tsx
@@ -1,10 +1,10 @@
 import { Block } from "expo-ui-kit";
 import React from "react";
 import { StyleSheet } from "react-native";
-import PopularComponent from "./PopularComponent";
 import GenresComponent from "./GenresComponent";
 import LegendaryMovieComponent from "./LegendaryMovieComponent";
 import CastAndCrewComponent from "./CastAndCrewComponent";
+import PopularComponent from "./PopularComponent";
 
 export default function BodyComponent() {
   return (
