+++ screens/Home/components/body/CastAndCrewComponent.tsx
@@ -1,22 +1,22 @@
 import { Block, Text } from "expo-ui-kit";
 import React from "react";
 import { StyleSheet } from "react-native";
-import Colors from "../../../models/constants/Colors";
-import PopularListComponent from "./PopularListComponent";
+import Colors from "../../../../models/constants/Colors";
+import CastAndCrewListComponent from "./CastAndCrewListComponent";
 
-export default function LegendaryMovieComponent() {
+export default function CastAndCrewComponent() {
   return (
     <Block style={styles.container}>
       <Block row space="between">
         <Text size={18} color={Colors.basicTextColor}>
-          Legendary Movies
+          Cast & Crew
         </Text>
         <Text size={16} color={Colors.basicLinkColor}>
           See all
         </Text>
       </Block>
 
-      <PopularListComponent />
+      <CastAndCrewListComponent />
     </Block>
   );
 }
@@ -24,5 +24,6 @@ export default function LegendaryMovieComponent() {
 const styles = StyleSheet.create({
   container: {
     padding: 20,
+    marginBottom:70
   },
 });
