+++ screens/Home/components/body/LegendaryMovieComponent.tsx
@@ -1,22 +1,22 @@
 import { Block, Text } from "expo-ui-kit";
 import React from "react";
 import { StyleSheet } from "react-native";
-import Colors from "../../../models/constants/Colors";
-import CastAndCrewListComponent from "./CastAndCrewListComponent";
+import Colors from "../../../../models/constants/Colors";
+import PopularListComponent from "./PopularListComponent";
 
-export default function CastAndCrewComponent() {
+export default function LegendaryMovieComponent() {
   return (
     <Block style={styles.container}>
       <Block row space="between">
         <Text size={18} color={Colors.basicTextColor}>
-          Cast & Crew
+          Legendary Movies
         </Text>
         <Text size={16} color={Colors.basicLinkColor}>
           See all
         </Text>
       </Block>
 
-      <CastAndCrewListComponent />
+      <PopularListComponent />
     </Block>
   );
 }
@@ -24,6 +24,5 @@ export default function CastAndCrewComponent() {
 const styles = StyleSheet.create({
   container: {
     padding: 20,
-    marginBottom:70
   },
 });
